+++ API/src/app/Categoty/index.ts
@@ -1,0 +1,5 @@
+export * from './controllers/category.controller';
+export * from './entities/category.entity';
+export * from './repositories/category.repository';
+export * from './services/category.service';
+export * from './category.module';
