+++ API/src/app/DataPoint/index.ts
@@ -1,0 +1,4 @@
+export * from './controllers/dataPpoint.controller';
+export * from './entities/dataPoint.entity';
+export * from './repositories/dataPoint.repository';
+export * from './services/dataPoint.service';
