+++ API/src/app/DataPoint/repositories/dataPoint.repository.ts
@@ -1,0 +1,1 @@
+export class DataPointRepository {}
