--- API/src/app/DatasetEntries/datasetentries.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import DataSetEntriesController from './datasetentries.controller';
-import { DataSetEntriesService } from './datasetentries.service';
-
-@Module({
-    imports: [],
-    controllers: [DataSetEntriesController],
-    providers: [DataSetEntriesService],
-})
-export class DataSetEntriesModule {}
