--- API/src/app/DatasetEntries/datasetentries.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { DataSetEntriesAPI } from './datasetentries.models';
-
-@Injectable()
-export class DataSetEntriesService {
-    getData(entriesQuerry: DataSetEntriesAPI) {
-        return entriesQuerry;
-    }
-}
