+++ API/src/app/DatasetEntry/datasetEntry.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@nestjs/common';
+import { DataSetEntry } from './datasetEntry.models';
+
+@Injectable()
+export class DataSetEntryService {
+    getData(entriesQuery: DataSetEntry) {
+        return entriesQuery;
+    }
+}
