+++ API/src/app/DatasetEntry/index.ts
@@ -3,4 +3,4 @@
  * @desc        Module API Endpoint for DataSetEntries
  */
 
-export * from './datasetentries.module';
+export * from './datasetEntry.module';
