+++ API/src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { DatabaseType, getMetadataArgsStorage } from 'typeorm';
-import { CategoryModule } from './Categoty/category.module';
+import { CategoryModule } from './Categoty';
 import { DataPointModule } from './DataPoint/dataPoint.module';
 import { DSLModule } from './DSL/DSL.module';
 const mongodbConnection: DatabaseType = 'mongodb';
